requirements.txt
streamlit
supabase
python-dotenv
openai
plotly
pandas

